refactor: react query 리팩토링
